+++ shoes/canvas.c
@@ -7,7 +7,7 @@
 #include "shoes/canvas.h"
 #include "shoes/ruby.h"
 #include "shoes/world.h"
-#include "shoes/native.h"
+#include "shoes/native/native.h"
 #include "shoes/http.h"
 
 #define SETUP_SHAPE() \
