+++ shoes/http/rbload.c
@@ -11,7 +11,7 @@
 #include "shoes/http.h"
 #include "shoes/version.h"
 #include "shoes/world.h"
-#include "shoes/native.h"
+#include "shoes/native/native.h"
 #include "shoes/canvas.h"
 
 void shoes_download(shoes_http_request *req) {
