+++ shoes/native/gtk/gtkcomboboxtextalt.c
@@ -4,7 +4,7 @@
 #include "shoes/ruby.h"
 #include "shoes/config.h"
 #include "shoes/world.h"
-#include "shoes/native.h"
+#include "shoes/native/native.h"
 #include "shoes/internal.h"
 
 #include "gtkcomboboxtextalt.h"
