+++ shoes/native/gtk/gtkeditline.c
@@ -2,7 +2,7 @@
 #include "shoes/ruby.h"
 #include "shoes/config.h"
 #include "shoes/world.h"
-#include "shoes/native.h"
+#include "shoes/native/native.h"
 #include "shoes/internal.h"
 
 #include "shoes/native/gtk/gtkentryalt.h"
